Add an option to increase LSM6DS3 I2C speed to 700kHz
The maximum I2C speed stated in the datasheet is 400 kHz. However, it
also states maximum SPI speeds are much higher, so it's quite safe to
expect 700 kHz I2C will work as well, as most of the circuitry should be
shared.

However, the data lines on the ESCs were not made for such high speeds,
in particular it seems some controllers have 10k pullups on the SDA
line, which slows it down to borderline response for 700 kHz signal.
Note this depends a lot on the PCB trace capacitance of the particular
ESC.

Feature: Add an option to increase the LSM6DS3 communication speed to 700kHz
 >
 700 kHz speed was tested on the following controllers (single unit
 each):
 >
 - Tronic 250R (riding for ~2 months)
 >
 - JetFleet F6 (riding for ~5 months)
 >
 - Thor 300 (benchtop testing)
 >
 - Thor 301 (brief benchtop testing)
 >
 While no instability has been observed, there are reasons to believe
 the higher speed may not be reliable and more testing is needed. Until
 it's been proven reliable, great care should be taken when configuring
 the higher speed to verify it's stable for the particular controller
 unit.
 >
 The higher speed can be configured by setting the "Use magnetometer"
 IMU config option to _**False**_ (false = fast). This is because the
 default value of the option is _True_ and we want the slow speed by
 default.
 >
 The speed is printed to the VESC terminal after writing the IMU
 configuration and can be verified there.